Refactor LottieGen to enable parallel codegen, and enable parallel codegen.
This will make testing faster.
